pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "deep-strat"
version = "0.1.0"
description = "A knowledge-based agent system"
readme = "README.md"
requires-python = ">=3.10,<3.11"
license = "MIT"
authors = [
    { name = "Your Name", email = "your.email@example.com" }
]
dependencies = [
    "google-generativeai>=0.3.0",
    "beautifulsoup4>=4.12.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "SQLAlchemy>=2.0.0",
    "schedule>=1.2.0",
    "Flask>=2.0.0",
    "flask-cors",
    "google-genai",
    "langchain",
    "openai==1.69.0",
    "chromadb>=0.5.0,<0.7.0", 
    "langchain-community",
    "langchain-chroma",
    "langchain-openai",
    "langchain-voyageai",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "pre-commit>=3.0.0",
]

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 88

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=deep_strat --cov-report=term-missing" 